requirements.txt
dnspython==1.16.0
docutils==0.14
entrypoints==0.3
folium==0.10.1
geographiclib==1.50
geopy==1.20.0
jsonify==0.5
panda==0.3.1
pandas==0.25.3
pymongo==3.10.1
python-dateutil==2.8.0
tweepy==3.7.0
tweet-preprocessor==0.4.0
twitter==1.18.0
twitter-text-python==1.1.1
uuid==1.30
